codecmm: CB2 Crash or green screen occurs when memory is insufficient. [1/1]
PD#SWPL-117836

Problem:
When the memory is insufficient,
codecmm forcibly releases the cache,
causing the address of the scatter mem pointed to
in the avbc header to become invalid,
and an address exception access occurs.

Solution:
codecmm does not release the cache when the memory is insufficient.

Verify:
T5M

Change-Id: Ie3068506a26d5e480fe8e089e7bdb09b5fd45477
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>